+++ 01_Nbody/lib/Integrator.hpp
@@ -6,7 +6,7 @@ class Integrator {
 
 public:
     virtual ~Integrator();
-    virtual std::vector<Body> integrate(const std::vector<Body>& bodies, const double maxTimeStep);
+    virtual std::vector<Body> integrate(const std::vector<Body>& bodies, const long double maxTimeStep);
 };
 
 #endif // INTEGRATOR_HPP
