+++ 01_Nbody/lib/RK4.hpp
@@ -3,6 +3,6 @@
 #include "Integrator.hpp"
 class RK4 : public Integrator {
     public:
-        std::vector<Body> integrate(const std::vector<Body>& bodies, const double maxTimeStep) override;
+        std::vector<Body> integrate(const std::vector<Body>& bodies, const double long maxTimeStep) override;
 };
 #endif // RK4_HPP
